+++ src/main/java/com/powsybl/openloadflow/network/impl/Networks.java
@@ -110,57 +110,10 @@ public final class Networks {
         return LfNetwork.load(network, new LfNetworkLoaderImpl(), parameters, reporter);
     }
 
-    static class LfNetworkListImpl implements LfNetworkList {
-
-        private final List<LfNetwork> list;
-
-        LfNetworkListImpl(List<LfNetwork> list) {
-            this.list = Objects.requireNonNull(list);
-        }
-
-        @Override
-        public List<LfNetwork> getList() {
-            return list;
-        }
-
-        @Override
-        public void close() {
-            // nothing to clean
-        }
-    }
-
-    static class LfNetworkListVariantCleanupImpl implements LfNetworkList {
-
-        private final Network network;
-
-        private final String workingVariantId;
-
-        private final String tmpVariantId;
-
-        private final List<LfNetwork> list;
-
-        LfNetworkListVariantCleanupImpl(Network network, String workingVariantId, String tmpVariantId, List<LfNetwork> list) {
-            this.network = Objects.requireNonNull(network);
-            this.workingVariantId = Objects.requireNonNull(workingVariantId);
-            this.tmpVariantId = Objects.requireNonNull(tmpVariantId);
-            this.list = Objects.requireNonNull(list);
-        }
-
-        public List<LfNetwork> getList() {
-            return list;
-        }
-
-        @Override
-        public void close() {
-            network.getVariantManager().removeVariant(tmpVariantId);
-            network.getVariantManager().setWorkingVariant(workingVariantId);
-        }
-    }
-
     public static LfNetworkList load(Network network, LfNetworkParameters networkParameters,
                                      Set<Switch> switchesToOpen, Set<Switch> switchesToClose, Reporter reporter) {
         if (switchesToOpen.isEmpty() && switchesToClose.isEmpty()) {
-            return new LfNetworkListImpl(load(network, networkParameters, reporter));
+            return new LfNetworkList(load(network, networkParameters, reporter));
         } else {
             String tmpVariantId = "olf-tmp-" + UUID.randomUUID();
             String workingVariantId = network.getVariantManager().getWorkingVariantId();
@@ -173,7 +126,8 @@ public final class Networks {
             switchesToClose.stream().filter(sw -> sw.getVoltageLevel().getTopologyKind() == TopologyKind.NODE_BREAKER)
                     .forEach(sw -> sw.setRetained(true));
             switchesToClose.forEach(sw -> sw.setOpen(false)); // in order to be present in the network.
-            return new LfNetworkListVariantCleanupImpl(network, workingVariantId, tmpVariantId, load(network, networkParameters, reporter));
+            return new LfNetworkList(load(network, networkParameters, reporter),
+                                     new LfNetworkList.VariantCleaner(network, workingVariantId, tmpVariantId));
         }
     }
 }
